refactor(web): deduplicate ledger-adapter to single source of truth (#1582)
Two copies of the ledger-record-to-transcript-item projection existed:
- `dashboard/src/lib/ledger-adapter.ts` (React dashboard, Vite build)
- `src/web-server/frontend/ledger-adapter.ts` (legacy vanilla TS, esbuild)

A schema change had to be applied to both or the React dashboard would
silently diverge from the legacy UI.

Fix: extract the canonical implementation to a new shared module at
`src/web-server/shared/ledger-adapter.ts` that merges the best of both:
- All Wave 1 kinds from the legacy adapter: `tool_activity`, `rate_limit`,
  `progress`, `plan_mode`
- `subagentIndex` deduplication for `subagent_lifecycle` events
- Richer label formatting for subagent cards (agentType-based)
- Token breakdown in `done` notices
- `tool_result` unmatched case rendered as a notice (not a new tool item)
- All types self-contained: `ToolCallItem`, `SubagentItem`, `ToolIndex`,
  `SubagentIndex`, `SseStreamFrame` exported directly (no view-model.ts dep)

Both adapters become thin re-export wrappers:
- `src/web-server/frontend/ledger-adapter.ts` → `../shared/ledger-adapter.js`
- `dashboard/src/lib/ledger-adapter.ts` → `../../../src/web-server/shared/ledger-adapter`

The Vite build can resolve relative imports outside `dashboard/` without
any alias changes. The esbuild pipeline also follows relative imports.

Side-effects: `render.ts` and `render-incremental.ts` now import
`TranscriptItem` from `./ledger-adapter.js` (the re-export) instead of
`./view-model.js`, which was the correct source anyway since app.ts
only ever populates items from the ledger adapter. The `BgJobItem` type
that `render.ts` referenced from view-model is replaced with an inline
`Extract<>` alias.

Gates: `pnpm lint` (both tsc projects) ✓, `pnpm test` (19655 tests) ✓,
`node scripts/build-web-ui.mjs` (legacy esbuild) ✓
